Manual amplitude selection
Added
set_amplitude
set_grand_amp
in the DataHandler class to handle manually setting individual
amplitudes. Selecting a filename in the top listbox will allow selection
of amplitudes for grand average AND all trials for that run. To set
averages independent of grand average select them in the lower listbox.